internal/worker/config_test.go
package worker

/* ***** BEGIN GPL LICENSE BLOCK *****
 *
 * Original Code Copyright (C) 2022 Blender Foundation.
 *
 * This file is part of Flamenco.
 *
 * Flamenco is free software: you can redistribute it and/or modify it under
 * the terms of the GNU General Public License as published by the Free Software
 * Foundation, either version 3 of the License, or (at your option) any later
 * version.
 *
 * Flamenco is distributed in the hope that it will be useful, but WITHOUT ANY
 * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
 * A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License along with
 * Flamenco.  If not, see <https://www.gnu.org/licenses/>.
 *
 * ***** END GPL LICENSE BLOCK ***** */

import (
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestParseURL(t *testing.T) {
	test := func(expected, input string) {
		actualURL, err := ParseURL(input)
		assert.Nil(t, err)
		assert.Equal(t, expected, actualURL.String())
	}

	test("http://jemoeder:1234", "jemoeder:1234")
	test("http://jemoeder/", "jemoeder")
	test("opjehoofd://jemoeder:4213/xxx", "opjehoofd://jemoeder:4213/xxx")
}
